+++ src/Config/Schema/Docs.hs
@@ -1,4 +1,4 @@
-{-# Language OverloadedStrings, GADTs, GeneralizedNewtypeDeriving #-}
+{-# Language RecursiveDo, OverloadedStrings, GADTs, GeneralizedNewtypeDeriving #-}
 
 {-|
 Module      : Config.Schema.Docs
@@ -36,13 +36,18 @@ module Config.Schema.Docs
   ( generateDocs
   ) where
 
+import           Control.Applicative (liftA2)
+import           Control.Monad (unless)
+import           Control.Monad.Trans.State.Strict (runState, get, put, State)
 import           Data.List (intersperse)
 import           Data.List.NonEmpty (NonEmpty((:|)))
 import           Data.Map (Map)
+import           Data.Monoid (Monoid(..))
 import qualified Data.Map as Map
+import           Data.Semigroup
 import           Data.Text (Text)
 import qualified Data.Text as Text
-import           Text.PrettyPrint (Doc, fsep, text, ($+$), (<>), (<+>), nest, empty, hsep)
+import           Text.PrettyPrint (Doc, fsep, text, ($+$), (<+>), nest, empty, hsep)
 
 import           Config.Schema.Spec
 
@@ -53,7 +58,7 @@ generateDocs spec = vcat' docLines
     sectionLines :: (Text, Doc) -> [Doc]
     sectionLines (name, fields) = [text "", txt name, nest 4 fields]
 
-    (topMap, topDoc) = runDocBuilder (valuesDoc spec)
+    (topDoc, topMap) = runDocBuilder (valuesDoc spec)
 
     docLines =
       case runValueSpecs_ (pure . SomeSpec) spec of
@@ -77,7 +82,7 @@ data SomeSpec where SomeSpec :: ValueSpec a -> SomeSpec
 -- | Compute the documentation for a list of sections, store the
 -- documentation in the sections map and return the name of the section.
 sectionsDoc :: Text -> SectionSpecs a -> DocBuilder Doc
-sectionsDoc l spec = emitDoc l . vcat' =<< runSections_ (fmap pure . sectionDoc) spec
+sectionsDoc l spec = emitDoc l (vcat' <$> runSections_ (fmap pure . sectionDoc) spec)
 
 
 -- | Compute the documentation lines for a single key-value pair.
@@ -88,7 +93,7 @@ sectionDoc s =
     OptSection name desc w -> aux empty      name desc <$> valuesDoc w
   where
     aux req name desc val =
-      txt name <> ":" <+> req <+> val $+$
+      (txt name <> ":") <+> req <+> val $+$
       if Text.null desc
         then empty
         else nest 4 (fsep (txt <$> Text.splitOn " " desc)) -- line wrap logic
@@ -116,7 +121,7 @@ valueDoc w =
     AtomSpec a       -> pure ("`" <> txt a <> "`")
     AnyAtomSpec      -> pure "atom"
     SectionSpecs l s -> sectionsDoc l s
-    NamedSpec    l s -> emitDoc l =<< valuesDoc s
+    NamedSpec    l s -> emitDoc l (valuesDoc s)
     CustomSpec l w'  -> (txt l                 <+>) <$> valuesDoc w'
     ListSpec ws      -> ("list of"             <+>) <$> valuesDoc ws
     AssocSpec ws     -> ("association list of" <+>) <$> valuesDoc ws
@@ -124,17 +129,39 @@ valueDoc w =
 
 -- | A writer-like type. A mapping of section names and documentation
 -- lines is accumulated.
-newtype DocBuilder a = DocBuilder { runDocBuilder :: (Map Text Doc, a) }
-  deriving (Functor, Applicative, Monad, Monoid, Show)
+newtype DocBuilder a = DocBuilder (State (Map Text Doc) a)
+  deriving (Functor, Applicative, Monad)
+
+runDocBuilder :: DocBuilder a -> (a, Map Text Doc)
+runDocBuilder (DocBuilder b) = runState b mempty
+
+-- | lifts underlying Semigroup instance
+instance Semigroup a => Semigroup (DocBuilder a) where
+  (<>) = liftA2 (<>)
+
+-- | lifts underlying Monoid instance
+instance (Semigroup a, Monoid a) => Monoid (DocBuilder a) where
+  mempty  = pure mempty
+  mappend = (<>)
 
 
 -- | Given a section name and section body, store the body
 -- in the map of sections and return the section name.
 emitDoc ::
-  Text {- ^ section name -} ->
-  Doc  {- ^ section body -} ->
+  Text           {- ^ section name -} ->
+  DocBuilder Doc {- ^ section body -} ->
   DocBuilder Doc
-emitDoc l xs = DocBuilder (Map.singleton l xs, txt l)
+emitDoc l (DocBuilder sub) = DocBuilder $
+  do m <- get
+     unless (Map.member l m) $
+       do rec put $! Map.insert l val m
+              val <- sub
+          return ()
+     return (txt l)
+  -- by using a recursively defined do block and
+  -- inserting the element /before/ executing the @sub@
+  -- action we ensure that @sub@ doesn't attempt to
+  -- also explore elements named @l@
 
 ------------------------------------------------------------------------